+++ packages/logic-core/src/data/solver/allSolvers.ts
@@ -2,6 +2,7 @@ import Solver from './solver.js';
 import UniversalSolver from './universal/universalSolver.js';
 import BacktrackSolver from './backtrack/backtrackSolver.js';
 import Z3Solver from './z3/z3Solver.js';
+import CspuzSolver from './cspuz/cspuzSolver.js';
 
 const allSolvers = new Map<string, Solver>();
 
@@ -9,6 +10,7 @@ function register(prototype: Solver) {
   allSolvers.set(prototype.id, prototype);
 }
 
+register(new CspuzSolver());
 register(new BacktrackSolver());
 register(new UniversalSolver());
 register(new Z3Solver());