+++ packages/logic-core/src/data/solver/z3/z3Solver.ts
@@ -20,7 +20,7 @@ export default class Z3Solver extends Solver {
   public readonly author = 'Lysine';
 
   public readonly description =
-    'Good for confirming that a solution is unique, especially for larger puzzles. It is otherwise slower than most solvers in small to medium-sized puzzles.';
+    '(Obsolete) A WebAssembly solver that supports a limited set of rules and symbols.';
 
   public readonly supportsCancellation = false;